+++ src/torchmetrics/detection/diou.py
@@ -17,10 +17,10 @@ from torch import Tensor
 
 from torchmetrics.detection.iou import IntersectionOverUnion
 from torchmetrics.functional.detection.diou import _diou_compute, _diou_update
-from torchmetrics.utilities.imports import _MATPLOTLIB_AVAILABLE, _TORCHVISION_GREATER_EQUAL_0_13
+from torchmetrics.utilities.imports import _MATPLOTLIB_AVAILABLE, _TORCHVISION_AVAILABLE
 from torchmetrics.utilities.plot import _AX_TYPE, _PLOT_OUT_TYPE
 
-if not _TORCHVISION_GREATER_EQUAL_0_13:
+if not _TORCHVISION_AVAILABLE:
     __doctest_skip__ = ["DistanceIntersectionOverUnion", "DistanceIntersectionOverUnion.plot"]
 elif not _MATPLOTLIB_AVAILABLE:
     __doctest_skip__ = ["DistanceIntersectionOverUnion.plot"]
@@ -110,10 +110,10 @@ class DistanceIntersectionOverUnion(IntersectionOverUnion):
         respect_labels: bool = True,
         **kwargs: Any,
     ) -> None:
-        if not _TORCHVISION_GREATER_EQUAL_0_13:
+        if not _TORCHVISION_AVAILABLE:
             raise ModuleNotFoundError(
-                f"Metric `{self._iou_type.upper()}` requires that `torchvision` version 0.13.0 or newer is installed."
-                " Please install with `pip install torchvision>=0.13` or `pip install torchmetrics[detection]`."
+                f"Metric `{self._iou_type.upper()}` requires that `torchvision` is installed."
+                " Please install with `pip install torchmetrics[detection]`."
             )
         super().__init__(box_format, iou_threshold, class_metrics, respect_labels, **kwargs)