+++ src/torchmetrics/functional/detection/iou.py
@@ -15,9 +15,9 @@ from typing import Optional
 
 import torch
 
-from torchmetrics.utilities.imports import _TORCHVISION_GREATER_EQUAL_0_8
+from torchmetrics.utilities.imports import _TORCHVISION_AVAILABLE
 
-if not _TORCHVISION_GREATER_EQUAL_0_8:
+if not _TORCHVISION_AVAILABLE:
     __doctest_skip__ = ["intersection_over_union"]
 
 
@@ -114,10 +114,10 @@ def intersection_over_union(
                 [0.0000, 0.0000, 0.5654]])
 
     """
-    if not _TORCHVISION_GREATER_EQUAL_0_8:
+    if not _TORCHVISION_AVAILABLE:
         raise ModuleNotFoundError(
-            f"`{intersection_over_union.__name__}` requires that `torchvision` version 0.8.0 or newer is installed."
-            " Please install with `pip install torchvision>=0.8` or `pip install torchmetrics[detection]`."
+            f"`{intersection_over_union.__name__}` requires that `torchvision` is installed."
+            " Please install with `pip install torchmetrics[detection]`."
         )
     iou = _iou_update(preds, target, iou_threshold, replacement_val)
     return _iou_compute(iou, aggregate)